+++ dandelion-upm/dandelion-upm-api/src/main/java/cn/icepear/dandelion/upm/api/domain/dto/TreeNode.java
@@ -13,8 +13,8 @@ import java.util.List;
  */
 @Data
 public class TreeNode<T> implements Serializable {
-	protected int id;
-	protected int parentId;
+	protected Long id;
+	protected Long parentId;
 	protected List<T> children = new ArrayList<T>();
 
 	public void add(T node) {
