+++ dandelion-upm/dandelion-upm-api/src/main/java/cn/icepear/dandelion/upm/api/domain/entity/SysDeptRole.java
@@ -22,10 +22,10 @@ public class SysDeptRole extends Model<SysDeptRole> {
 	/**
 	 * 角色ID
 	 */
-	private Integer roleId;
+	private Long roleId;
 	/**
 	 * 部门ID
 	 */
-	private Integer deptId;
+	private Long deptId;
 
 }
